requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
Faker==30.8.2
pytest==8.3.3
selenium==4.26.0

